requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
python-multipart>=0.0.9
numpy>=1.26
pillow>=10.2
opencv-python-headless>=4.9
scikit-image>=0.22
anthropic>=0.40
httpx>=0.27
